+++ src/android/de/rouvenkruse/cordova/plugins/cookieemperor/CookieEmperor.java
@@ -1,0 +1,143 @@
+package de.rouvenkruse.cordova.plugins.cookieemperor;
+
+import org.apache.cordova.CallbackContext;
+import org.apache.cordova.CordovaPlugin;
+import org.apache.cordova.PluginResult;
+
+import org.json.JSONObject;
+import org.json.JSONArray;
+import org.json.JSONException;
+
+import org.apache.http.cookie.Cookie;
+
+import java.net.HttpCookie;
+
+import android.webkit.CookieManager;
+
+public class CookieMaster extends CordovaPlugin {
+
+    public static final String ACTION_GET_COOKIE_VALUE = "getCookieValue";
+    public static final String ACTION_SET_COOKIE_VALUE = "setCookieValue";
+    public static final String ACTION_CLEAR_COOKIES = "clearCookies";
+
+    @Override
+    public boolean execute(String action, JSONArray args, final CallbackContext callbackContext) throws JSONException {
+        if (ACTION_GET_COOKIE_VALUE.equals(action)) {
+            return this.getCookie(args, callbackContext);
+        }
+        else if (ACTION_SET_COOKIE_VALUE.equals(action)) {
+            return this.setCookie(args, callbackContext);
+        }
+        else if (ACTION_CLEAR_COOKIES.equals(action)) {
+            return this.clearAllCookies();
+        }
+
+        callbackContext.error("Invalid action");
+        return false;
+
+    }
+
+    /**
+     * returns cookie under given key
+     * @param args
+     * @param callbackContext
+     * @return
+     */
+    private boolean getCookie(JSONArray args, CallbackContext callbackContext) {
+        final String url = args.getString(0);
+        final String cookieName = args.getString(1);
+
+        cordova
+                .getThreadPool()
+                .execute(new Runnable() {
+                    public void run() {
+                        try {
+                            CookieManager cookieManager = CookieManager.getInstance();
+                            String[] cookies = cookieManager.getCookie(url).split("; ");
+                            String cookieValue = "";
+
+                            for (int i = 0; i < cookies.length; i++) {
+                                if (cookies[i].contains(cookieName + "=")) {
+                                    cookieValue = cookies[i].split("=")[1].trim();
+                                    break;
+                                }
+                            }
+
+                            JSONObject json = null;
+
+                            if (cookieValue != "") {
+                                json = new JSONObject("{cookieValue:\"" + cookieValue + "\"}");
+                            }
+
+                            if (json != null) {
+                                PluginResult res = new PluginResult(PluginResult.Status.OK, json);
+                                callbackContext.sendPluginResult(res);
+                            }
+                            else {
+                                callbackContext.error("Cookie not found!");
+                            }
+                        }
+                        catch (Exception e) {
+                            callbackContext.error(e.getMessage());
+                        }
+                    }
+                });
+
+        return true;
+    }
+
+    /**
+     * sets cookie value under given key
+     * @param args
+     * @param callbackContext
+     * @return boolean
+     */
+    private boolean setCookie(JSONArray args, CallbackContext callbackContext) {
+        final String url = args.getString(0);
+        final String cookieName = args.getString(1);
+        final String cookieValue = args.getString(2);
+
+        cordova
+                .getThreadPool()
+                .execute(new Runnable() {
+                    public void run() {
+                        try {
+                            HttpCookie cookie = new HttpCookie(cookieName, cookieValue);
+
+                            String cookieString = cookie.toString().replace("\"", "");
+                            CookieManager cookieManager = CookieManager.getInstance();
+                            cookieManager.setCookie(url, cookieString);
+
+                            PluginResult res = new PluginResult(PluginResult.Status.OK, "Successfully added cookie");
+                            callbackContext.sendPluginResult(res);
+                        }
+                        catch (Exception e) {
+                            callbackContext.error(e.getMessage());
+                        }
+                    }
+                });
+
+        return true;
+    }
+
+    /**
+     * Clears all cookies
+     * @return boolean
+     */
+    private boolean clearAllCookies() {
+        CookieManager cookieManager = CookieManager.getInstance();
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP_MR1) {
+            cookieManager.removeAllCookies();
+            cookieManager.flush();
+        }
+        else {
+            cookieManager.removeAllCookie();
+            cookieManager.removeSessionCookie();
+        }
+
+        callbackContext.success();
+
+        return true;
+    }
+}
