+++ src/ios/CDVCookieEmperor.h
@@ -1,14 +1,6 @@
-//
-//  CDVCookieMaster.h
-//
-//
-//  Created by Kristian Hristov on 12/16/14.
-//
-//
-
 #import <Cordova/CDVPlugin.h>
 
-@interface CDVCookieMaster : CDVPlugin
+@interface CDVCookieEmperor : CDVPlugin
 
 - (void)getCookieValue:(CDVInvokedUrlCommand*)command;
 - (void)setCookieValue:(CDVInvokedUrlCommand*)command;
