--- www/cookieMaster.js
@@ -1,25 +1,0 @@
-var cookieMaster = {
-
-    getCookieValue: function(url, cookieName, successCallback, errorCallback) {
-        cordova.exec(successCallback,
-                    errorCallback,
-                    'CookieMaster', 'getCookieValue',
-                    [url, cookieName]
-        );
-    },
-    setCookieValue: function (url, cookieName, cookieValue, successCallback, errorCallback) {
-        cordova.exec(successCallback,
-                    errorCallback,
-                    'CookieMaster', 'setCookieValue',
-                    [url, cookieName, cookieValue]
-        );
-    },
-    clear: function(successCallback, errorCallback) {
-        cordova.exec(successCallback,
-                    errorCallback,
-                    'CookieMaster', 'clearCookies',
-                    []
-        );
-    }
-};
-module.exports = cookieMaster;
